requirements.txt
streamlit>=1.33.0
pymongo>=4.6.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
email-validator>=2.1.0
yfinance>=0.2.32
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
plotly>=5.17.0
scikit-learn>=1.3.0